+++ apps/extension/src/core/domains/app/handler.ts
@@ -15,6 +15,7 @@ import { getEthDerivationPath } from "@core/domains/ethereum/helpers"
 import { genericSubscription } from "@core/handlers/subscriptions"
 import { talismanAnalytics } from "@core/libs/Analytics"
 import { ExtensionHandler } from "@core/libs/Handler"
+import { windowManager } from "@core/libs/WindowManager"
 import { chaindataProvider } from "@core/rpcs/chaindata"
 import type { MessageTypes, RequestTypes, ResponseType } from "@core/types"
 import { Port } from "@core/types/base"
@@ -210,7 +211,7 @@ export default class AppHandler extends ExtensionHandler {
     this.stores.app.set({ onboarded: "FALSE" })
     await this.stores.password.reset()
     await this.stores.seedPhrase.clear()
-    await this.state.openOnboarding("/import?resetWallet=true")
+    await windowManager.openOnboarding("/import?resetWallet=true")
     // since all accounts are being wiped, all sites need to be reset - so they may as well be wiped.
     await this.stores.sites.clear()
 
@@ -219,7 +220,7 @@ export default class AppHandler extends ExtensionHandler {
 
   private async dashboardOpen({ route }: RequestRoute): Promise<boolean> {
     if (!(await this.stores.app.getIsOnboarded())) return this.onboardOpen()
-    this.state.openDashboard({ route })
+    windowManager.openDashboard({ route })
     return true
   }
 
@@ -231,7 +232,7 @@ export default class AppHandler extends ExtensionHandler {
       const params = new URLSearchParams()
       if (from) params.append("from", from)
       if (tokenId) params.append("tokenId", tokenId)
-      await this.state.popupOpen(`#/send?${params.toString()}`)
+      await windowManager.popupOpen(`#/send?${params.toString()}`)
     } else {
       // TODO : delete as soon as we remove the SEND_FUNDS_V2 feature flag
       let transferableTokenId: string | undefined = undefined
@@ -249,7 +250,7 @@ export default class AppHandler extends ExtensionHandler {
     const queryUrl = Browser.runtime.getURL("dashboard.html")
     const [tab] = await Browser.tabs.query({ url: queryUrl })
     if (!tab) {
-      await this.state.openDashboard({ route: "/portfolio" })
+      await windowManager.openDashboard({ route: "/portfolio" })
       // wait for newly created page to load and subscribe to backend (max 5 seconds)
       for (let i = 0; i < 50 && !this.#modalOpenRequest.observed; i++) await sleep(100)
     }
@@ -257,7 +258,7 @@ export default class AppHandler extends ExtensionHandler {
   }
 
   private onboardOpen(): boolean {
-    this.state.openOnboarding()
+    windowManager.openOnboarding()
     return true
   }
 
@@ -267,7 +268,7 @@ export default class AppHandler extends ExtensionHandler {
   }
 
   private promptLogin(closeOnSuccess: boolean): boolean {
-    this.state.promptLogin(closeOnSuccess)
+    windowManager.popupOpen(`?closeOnSuccess=${closeOnSuccess}`)
     return true
   }