+++ apps/extension/src/ui/apps/popup/pages/Sign/substrate.tsx
@@ -1,5 +1,5 @@
 import { AccountJsonHardwareSubstrate } from "@core/domains/accounts/types"
-import { SignerPayloadRaw, SigningRequest } from "@core/domains/signing/types"
+import { KnownSigningRequestIdOnly, SignerPayloadRaw } from "@core/domains/signing/types"
 import { SimpleButton } from "@talisman/components/SimpleButton"
 import { Content, Footer, Header } from "@ui/apps/popup/Layout"
 import { AccountPill } from "@ui/domains/Account/AccountPill"
@@ -19,8 +19,8 @@ import { Container } from "./common"
 const LedgerSubstrate = lazy(() => import("@ui/domains/Sign/LedgerSubstrate"))
 
 export const SubstrateSignRequest = () => {
-  const { id } = useParams() as { id: string }
-  const signingRequest = useSigningRequestById(id) as SigningRequest | undefined
+  const { id } = useParams() as KnownSigningRequestIdOnly<"substrate-sign">
+  const signingRequest = useSigningRequestById(id)
   const { url, request, approve, reject, status, message, account, chain, approveHardware } =
     usePolkadotSigningRequest(signingRequest)
   const { analysing, txDetails, error: txDetailsError } = usePolkadotTransactionDetails(id)
@@ -92,10 +92,10 @@ export const SubstrateSignRequest = () => {
                 </SimpleButton>
               </div>
             )}
-            {account.isHardware && (
+            {account.isHardware && reject && (
               <Suspense fallback={null}>
                 <LedgerSubstrate
-                  payload={(request as SigningRequest["request"]).payload}
+                  payload={request.payload}
                   account={account as AccountJsonHardwareSubstrate}
                   genesisHash={chain?.genesisHash ?? account?.genesisHash ?? undefined}
                   onSignature={approveHardware}