+++ apps/extension/src/ui/hooks/useSigningRequestById.ts
@@ -1,22 +1,29 @@
-import type { AnySigningRequest } from "@core/domains/signing/types"
+import type {
+  AnySigningRequest,
+  AnySigningRequestID,
+  SigningRequestID,
+  SigningRequests,
+} from "@core/domains/signing/types"
 import { api } from "@ui/api"
 import { useCallback } from "react"
 import { BehaviorSubject } from "rxjs"
 
 import { useMessageSubscription } from "./useMessageSubscription"
 
-const INITIAL_SUBJECT_VALUE: Record<string, AnySigningRequest> = {}
+const INITIAL_SUBJECT_VALUE: Record<AnySigningRequestID, AnySigningRequest> = {}
 
 // public hook
-export const useSigningRequestById = (id: string): AnySigningRequest | undefined => {
+export const useSigningRequestById = <T extends keyof SigningRequests>(
+  id: SigningRequestID<T>
+): SigningRequests[T][0] | undefined => {
   const subscribe = useCallback(
-    (signingRequest: BehaviorSubject<Record<string, AnySigningRequest>>) =>
+    (signingRequest: BehaviorSubject<Record<AnySigningRequestID, AnySigningRequest>>) =>
       api.subscribeSigningRequest(id, (v) => signingRequest.next({ [id]: v })),
     [id]
   )
 
   const transform = useCallback(
-    (signingRequests: Record<string, AnySigningRequest>) => signingRequests[id],
+    (signingRequests: Record<AnySigningRequestID, AnySigningRequest>) => signingRequests[id],
     [id]
   )