+++ packages/util/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @talismn/util
 
+## 0.1.6
+
+### Patch Changes
+
+- 8adc7f06: feat: switched build tool to preconstruct
+
 ## 0.1.5
 
 ### Patch Changes